riscv: Add ISA double trap extension testing
This test allows to test the double trap implementation of hardware as
well as the SBI FWFT and SSE support for double trap. The tests will try
to trigger double trap using various sequences and will test to receive
the SSE double trap event if supported.

It is provided as a separate tests of the SBI one for two reasons:
- It isn't specifically testing SBI per se.
- It ends up by trying to crash into in M-mode.

Signed-off-by: Clément Léger <cleger@rivosinc.com>